fix(W5): split Paparazzi hi-locale tests into separate classes (one rule per class)
Two @get:Rule Paparazzi instances in one class cause layoutlib to initialize
twice per test, resulting in IllegalStateException on CI (and Windows).

Fix: move all paparazziHi @Rule + _hi tests out of the en-locale classes into
new dedicated *HiPaparazziTest classes — each with exactly one Paparazzi rule:
  - AuthScreenHiPaparazziTest (locale="hi")
  - ActiveJobScreenHiPaparazziTest (locale="hi")
  - JobOfferScreenHiPaparazziTest (locale="hi")
  - EarningsScreenHiPaparazziTest (locale="hi")

Existing en-locale test classes are restored to single-rule form.
