+++ technician-app/app/src/test/kotlin/com/homeservices/technician/ui/activeJob/ActiveJobScreenHiPaparazziTest.kt
@@ -9,11 +9,11 @@ import com.homeservices.technician.domain.activeJob.model.LatLng
 import org.junit.Rule
 import org.junit.Test
 
-public class ActiveJobScreenPaparazziTest {
+public class ActiveJobScreenHiPaparazziTest {
     @get:Rule
     public val paparazzi: Paparazzi =
         Paparazzi(
-            deviceConfig = DeviceConfig.PIXEL_5,
+            deviceConfig = DeviceConfig.PIXEL_5.copy(locale = "hi"),
             theme = "android:Theme.Material3.DayNight.NoActionBar",
         )
 
@@ -30,8 +30,10 @@ public class ActiveJobScreenPaparazziTest {
             slotWindow = "10:00-12:00",
         )
 
+    // Goldens recorded on CI Linux via paparazzi-record.yml; never locally on Windows.
+
     @Test
-    public fun activeJobScreen_enRoute(): Unit {
+    public fun activeJobScreen_enRoute_hi(): Unit {
         paparazzi.snapshot {
             HomeservicesTheme(darkTheme = false) {
                 ActiveJobScreenContent(
@@ -46,22 +48,7 @@ public class ActiveJobScreenPaparazziTest {
     }
 
     @Test
-    public fun activeJobScreen_reached(): Unit {
-        paparazzi.snapshot {
-            HomeservicesTheme(darkTheme = false) {
-                ActiveJobScreenContent(
-                    uiState = ActiveJobUiState.Active(aJob(ActiveJobStatus.REACHED), ActiveJobAction.START_WORK),
-                    onTransitionRequested = {},
-                    onPhotoCancelled = {},
-                    onPhotoConfirmed = {},
-                    onPhotoRetake = {},
-                )
-            }
-        }
-    }
-
-    @Test
-    public fun activeJobScreen_inProgress(): Unit {
+    public fun activeJobScreen_inProgress_hi(): Unit {
         paparazzi.snapshot {
             HomeservicesTheme(darkTheme = false) {
                 ActiveJobScreenContent(
@@ -74,5 +61,4 @@ public class ActiveJobScreenPaparazziTest {
             }
         }
     }
-
 }
